fix bug in stats field
